Fixed a few typos and suggested a few improvements
There were a few places where monit and mmonit were mixed up, and I think the monitrc file needs to be "600" permissions or monit refuses to start.

I think the nodeup command was checking the default host (one of the mainnet nodes).  I've adjusted it so that it checks the localhost, which seems more useful.  (I suspect nodeup originally defaulted to checking localhost.)