Hardhat boilerplate documentation
v1.0 of Hardhat documentation for Edgeware. This is currently a boilerplate for Hardhat deployment borrowed from [Moonbeam](https://docs.moonbeam.network/integrations/hardhat/) and tutorial documentation and resources from [Hardhat](https://hardhat.org/getting-started/). The current state of this documentation is not fully functional due to the current state of the Beresheet testnet and some missing features such as the discord bot faucet. Please make pull requests as needed.